services/web-be/requirements.txt
fastapi[all]
pydantic
pymongo[srv]
python-dotenv
pytest
